split shard messenger to new module
