+++ translations/pt.json
@@ -5,7 +5,7 @@
     "🟢 Great": "🟢 Ótimo",
     "⭐ Legendary ⭐": "⭐ Lendário ⭐",
     "🔴 Poor": "🔴 Ruim",
-    "🛒 {{trader.name}} restock in {{numMinutes}} minutes 🛒": "🛒 {{trader.name}} renova o estoque em {{numMinutes}} minutos 🛒",
+    "🛒 {{traderName}} restock in {{numMinutes}} minutes 🛒": "🛒 {{traderName}} renova o estoque em {{numMinutes}} minutos 🛒",
     "❌ ERROR ❌": "❌ ERRO ❌",
     "{{thingName}} set to {{level}}.": "{{thingName}} teve o nível definido para {{level}}.",
     "A Dmg": "Dano de A",
